# internal/databroker: make Sync send data in smaller batches